backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
click==8.3.1
dnspython==2.8.0
Flask==3.1.3
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.11.0
pymongo==4.16.0
python-dotenv==1.2.2
Werkzeug==3.1.6
Flask-Limiter==3.5.0
